requirements.txt
# 深度学习框架
torch>=1.12.0
torchvision>=0.13.0
torch-geometric>=2.0.0

# 数据处理
pandas>=1.4.0
numpy>=1.21.0
scipy>=1.8.0
scikit-learn>=1.1.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.8.0
folium>=0.12.0
geopandas>=0.10.0

# 配置管理
pyyaml>=6.0
hydra-core>=1.1.0

# 进度条和日志
tqdm>=4.64.0
loguru>=0.6.0

# 测试框架
pytest>=7.0.0
pytest-cov>=3.0.0

# 文档生成
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# 数据验证
pydantic>=1.9.0

# 实验管理
mlflow>=1.26.0
wandb>=0.12.0